+++ app/src/androidTest/java/com/codepath/randompet/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.driuft.random_pets_starter
+package com.codepath.randompet
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.driuft.random_pets_starter", appContext.packageName)
+        assertEquals("com.codepath.randompet", appContext.packageName)
     }
 }
